Move proactive copy to readonly session
