+++ Public/Src/Cache/ContentStore/Distributed/Sessions/DistributedContentSession.cs
@@ -2,12 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using BuildXL.Cache.ContentStore.Distributed.NuCache;
 using BuildXL.Cache.ContentStore.Distributed.Stores;
 using BuildXL.Cache.ContentStore.Hashing;
 using BuildXL.Cache.ContentStore.Interfaces.Extensions;
@@ -18,8 +15,6 @@ using BuildXL.Cache.ContentStore.Sessions.Internal;
 using BuildXL.Cache.ContentStore.Tracing;
 using BuildXL.Cache.ContentStore.Tracing.Internal;
 using BuildXL.Cache.ContentStore.UtilitiesCore;
-using BuildXL.Cache.ContentStore.Utils;
-using BuildXL.Utilities.Collections;
 using BuildXL.Utilities.Tracing;
 
 namespace BuildXL.Cache.ContentStore.Distributed.Sessions
@@ -31,21 +26,7 @@ namespace BuildXL.Cache.ContentStore.Distributed.Sessions
     public class DistributedContentSession<T> : ReadOnlyDistributedContentSession<T>, IContentSession
         where T : PathBase
     {
-        private enum Counters
-        {
-            GetLocationsSatisfiedFromLocal,
-            GetLocationsSatisfiedFromRemote
-        }
-
-        private readonly CounterCollection<Counters> _counters = new CounterCollection<Counters>();
         private readonly SemaphoreSlim _putFileGate;
-
-        private RocksDbContentPlacementPredictionStore _predictionStore;
-        private string _buildId = null;
-        private ContentHash? _buildIdHash = null;
-        private readonly ConcurrentBigSet<ContentHash> _pendingProactivePuts = new ConcurrentBigSet<ContentHash>();
-
-        private static readonly string PredictionBlobNameFile = "blobName.txt";
 
         /// <summary>
         /// Initializes a new instance of the <see cref="DistributedContentSession{T}"/> class.
@@ -72,56 +53,6 @@ namespace BuildXL.Cache.ContentStore.Distributed.Sessions
                 settings)
         {
             _putFileGate = new SemaphoreSlim(settings.MaximumConcurrentPutFileOperations);
-        }
-
-        /// <inheritdoc />
-        protected override async Task<BoolResult> StartupCoreAsync(OperationContext context)
-        {
-            await base.StartupCoreAsync(context).ThrowIfFailure();
-            if (Constants.TryExtractBuildId(Name, out _buildId) && Guid.TryParse(_buildId, out var buildIdGuid))
-            {
-                // Generate a fake hash for the build and register a content entry in the location store to represent
-                // machines in the build ring
-                _buildIdHash = new ContentHash(HashType.MD5, buildIdGuid.ToByteArray());
-
-                Tracer.Info(context, $"Registering machine with build {_buildId} (build id hash: {_buildIdHash.Value.ToShortString()}");
-                await ContentLocationStore.RegisterLocalLocationAsync(context, new[] { new ContentHashWithSize(_buildIdHash.Value, _buildId.Length) }, context.Token, UrgencyHint.Nominal).ThrowIfFailure();
-            }
-
-            await InitializePredictionStoreAsync(context);
-
-            return BoolResult.Success;
-        }
-
-        private Task InitializePredictionStoreAsync(OperationContext context)
-        {
-            return context.PerformOperationAsync(
-                Tracer,
-                async () =>
-                {
-                    var centralStorage = (ContentLocationStore as TransitioningContentLocationStore)?.LocalLocationStore?.CentralStorage;
-
-                    if (Settings.ContentPlacementPredictionsBlob != null)
-                    {
-                        var checkpointDirectory = Path.Combine(LocalCacheRootMachineLocation.Path, "PlacementPredictions");
-                        _predictionStore = new RocksDbContentPlacementPredictionStore(checkpointDirectory, clean: false);
-                        await _predictionStore.StartupAsync(context).ThrowIfFailure();
-
-                        var fileName = Path.Combine(checkpointDirectory, PredictionBlobNameFile);
-                        if (!File.Exists(fileName) || File.ReadAllText(fileName) != Settings.ContentPlacementPredictionsBlob)
-                        {
-                            Directory.Delete(checkpointDirectory);
-
-                            Directory.CreateDirectory(checkpointDirectory);
-
-                            var zipFile = Path.Combine(checkpointDirectory, "snapshot.zip");
-                            await centralStorage.TryGetFileAsync(context, Settings.ContentPlacementPredictionsBlob, new AbsolutePath(zipFile)).ThrowIfFailure();
-                            _predictionStore.UncompressSnapshot(context, zipFile).ThrowIfFailure();
-                        }
-                    }
-
-                    return BoolResult.Success;
-                });
         }
 
         /// <inheritdoc />
@@ -307,150 +238,5 @@ namespace BuildXL.Cache.ContentStore.Distributed.Sessions
 
             return putResult;
         }
-
-        private Task<ProactiveCopyResult> ProactiveCopyIfNeededAsync(OperationContext context, ContentHash hash, string path = null)
-        {
-            if (!_pendingProactivePuts.Add(hash))
-            {
-                return Task.FromResult(ProactiveCopyResult.CopyNotRequiredResult);
-            }
-
-            return context.PerformOperationAsync(
-                Tracer,
-                traceErrorsOnly: true,
-                operation: async () =>
-                {
-                    try
-                    {
-                        var hashArray = _buildIdHash != null
-                            ? new[] { hash, _buildIdHash.Value }
-                            : new[] { hash };
-
-                        // First check in local location store, then global if failed.
-                        var getLocationsResult = await ContentLocationStore.GetBulkAsync(context, hashArray, context.Token, UrgencyHint.Nominal, GetBulkOrigin.Local);
-                        if (getLocationsResult.Succeeded && getLocationsResult.ContentHashesInfo[0].Locations.Count > Settings.ProactiveCopyLocationsThreshold)
-                        {
-                            _counters[Counters.GetLocationsSatisfiedFromLocal].Increment();
-                            return ProactiveCopyResult.CopyNotRequiredResult;
-                        }
-                        else
-                        {
-                            getLocationsResult += await ContentLocationStore.GetBulkAsync(context, hashArray, context.Token, UrgencyHint.Nominal, GetBulkOrigin.Global).ThrowIfFailure();
-                            _counters[Counters.GetLocationsSatisfiedFromRemote].Increment();
-                        }
-
-                        if (getLocationsResult.ContentHashesInfo[0].Locations.Count > Settings.ProactiveCopyLocationsThreshold)
-                        {
-                            return ProactiveCopyResult.CopyNotRequiredResult;
-                        }
-
-                        IReadOnlyList<MachineLocation> buildRingMachines = null;
-
-                        // Get random machine inside build ring
-                        Task<BoolResult> insideRingCopyTask;
-                        if ((Settings.ProactiveCopyMode & ProactiveCopyMode.InsideRing) != 0)
-                        {
-                            if (_buildIdHash != null)
-                            {
-                                buildRingMachines = getLocationsResult.ContentHashesInfo[getLocationsResult.ContentHashesInfo.Count - 1].Locations;
-                                var candidates = buildRingMachines
-                                    .Where(m => !m.Equals(LocalCacheRootMachineLocation))
-                                    .Where(m => ContentLocationStore.IsMachineActive(m)).ToArray();
-
-                                if (candidates.Length > 0)
-                                {
-                                    var candidate = candidates[ThreadSafeRandom.Generator.Next(0, candidates.Length)];
-                                    insideRingCopyTask = RequestOrPushContentAsync(context, hash, candidate, isInsideRing: true);
-                                }
-                                else
-                                {
-                                    insideRingCopyTask = Task.FromResult(new BoolResult("Could not find any machines belonging to the build ring."));
-                                }
-                            }
-                            else
-                            {
-                                insideRingCopyTask = Task.FromResult(new BoolResult("BuildId was not specified, so machines in the build ring cannot be found."));
-                            }
-                        }
-                        else
-                        {
-                            insideRingCopyTask = BoolResult.SuccessTask;
-                        }
-
-                        buildRingMachines ??= new[] { LocalCacheRootMachineLocation };
-
-                        Task<BoolResult> outsideRingCopyTask;
-                        if ((Settings.ProactiveCopyMode & ProactiveCopyMode.OutsideRing) != 0)
-                        {
-                            Result<MachineLocation> getLocationResult = null;
-                            if (_predictionStore != null && path != null)
-                            {
-                                var machines = _predictionStore.GetTargetMachines(context, path);
-                                if (machines?.Count > 0)
-                                {
-                                    var index = ThreadSafeRandom.Generator.Next(0, machines.Count);
-                                    getLocationResult = new Result<MachineLocation>(new MachineLocation(machines[index]));
-                                }
-                            }
-
-                            if (getLocationResult == null)
-                            {
-                                getLocationResult = ContentLocationStore.GetRandomMachineLocation(except: buildRingMachines);
-                            }
-
-                            if (getLocationResult.Succeeded)
-                            {
-                                var candidate = getLocationResult.Value;
-                                outsideRingCopyTask = RequestOrPushContentAsync(context, hash, candidate, isInsideRing: false);
-                            }
-                            else
-                            {
-                                outsideRingCopyTask = Task.FromResult(new BoolResult(getLocationResult));
-                            }
-                        }
-                        else
-                        {
-                            outsideRingCopyTask = BoolResult.SuccessTask;
-                        }
-
-                        return new ProactiveCopyResult(await insideRingCopyTask, await outsideRingCopyTask);
-                    }
-                    finally
-                    {
-                        _pendingProactivePuts.Remove(hash);
-                    }
-                });
-        }
-
-        private async Task<BoolResult> RequestOrPushContentAsync(OperationContext context, ContentHash hash, MachineLocation target, bool isInsideRing)
-        {
-            if (Settings.PushProactiveCopies)
-            {
-                return await DistributedCopier.PushFileAsync(
-                    context,
-                    hash,
-                    target,
-                    async () =>
-                    {
-                        var streamResult = await Inner.OpenStreamAsync(context, hash, context.Token);
-                        if (streamResult.Succeeded)
-                        {
-                            return streamResult.Stream;
-                        }
-
-                        return null;
-                    }
-                    , isInsideRing);
-            }
-            else
-            {
-                return await DistributedCopier.RequestCopyFileAsync(context, hash, target, isInsideRing);
-            }
-        }
-
-        /// <inheritdoc />
-        protected override CounterSet GetCounters() =>
-            base.GetCounters()
-                .Merge(_counters.ToCounterSet());
     }
 }
